BasantBookFestival/requirements.txt
asgiref==3.2.7
django-filter==2.2.0
django-phonenumber-field==4.0.0
django-phonenumbers==1.0.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.5
phonenumbers==8.12.4
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
google==2.0.3
google-api-core==1.16.0
google-auth==1.12.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
Pillow==7.2.0
django-cors-headers==3.5.0
sentry-sdk==0.17.3
